requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.6
distro==1.9.0
frozenlist==1.4.1
greenlet==3.0.3
groq==0.8.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.2
langchain-groq==0.1.4
langchain-openai==0.1.8
langchain-pinecone==0.1.1
langchain-text-splitters==0.2.0
langchainhub==0.1.17
langsmith==0.1.64
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.30.5
orjson==3.10.3
packaging==23.2
pinecone-client==3.2.2
pydantic==2.7.2
pydantic_core==2.18.3
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
sniffio==1.3.1
python-dotenv
soupsieve==2.5
SQLAlchemy==2.0.30
tenacity==8.3.0
tiktoken==0.7.0
flask
tqdm==4.66.4
types-requests==2.32.0.20240523
typing-inspect==0.9.0
typing_extensions==4.12.0
urllib3==2.2.1
yarl==1.9.4